dra: only move if file does exist (#11477)

+++ .ci/scripts/prepare-release-manager.sh
@@ -23,10 +23,14 @@ cp build/distributions/dependencies.csv \
 #       or the unified release process the one to do the transformation
 for i in build/distributions/*linux-arm64.docker.tar.gz*
 do
+  if [ -e "$i" ] ; then
     mv "$i" "${i/linux-arm64.docker.tar.gz/docker-image-arm64.tar.gz}"
+  fi
 done
 
 for i in build/distributions/*linux-amd64.docker.tar.gz*
 do
+  if [ -e "$i" ] ; then
     mv "$i" "${i/linux-amd64.docker.tar.gz/docker-image.tar.gz}"
+  fi
 done
